Create better(?) indexes

+++ uiDivStats.sh
@@ -816,7 +816,7 @@ Trim_DNS_DB(){
 }
 
 Process_Upgrade(){
-	if [ ! -f "$SCRIPT_DIR/.upgraded" ]; then
+	if [ ! -f "$SCRIPT_DIR/.upgraded" ] && [ ! -f "$SCRIPT_DIR/.upgraded2" ]; then
 		opkg update
 		opkg install grep
 		opkg install sqlite3-cli
@@ -826,8 +826,9 @@ Process_Upgrade(){
 		{
 			echo "PRAGMA journal_mode=WAL;"
 			echo "CREATE TABLE IF NOT EXISTS [dnsqueries] ([QueryID] INTEGER PRIMARY KEY NOT NULL, [Timestamp] NUMERIC NOT NULL, [SrcIP] TEXT NOT NULL,[ReqDmn] TEXT NOT NULL,[QryType] Text NOT NULL,[Result] Text NOT NULL);"
-			echo "create index idx_dns on dnsqueries (Timestamp,ReqDmn,Result);"
-			echo "create index idx_dns_clients on dnsqueries (SrcIP,Timestamp,ReqDmn,Result);"
+			echo "create index idx_dns_domains on dnsqueries (ReqDmn,Timestamp);"
+			echo "create index idx_dns_time on dnsqueries (Timestamp,ReqDmn);"
+			echo "create index idx_dns_clients on dnsqueries (SrcIP,Timestamp,ReqDmn);"
 		}  > /tmp/uidivstats.sql
 		while ! "$SQLITE3_PATH" "$DNS_DB" < /tmp/uidivstats.sql >/dev/null 2>&1; do
 			sleep 1
@@ -835,6 +836,22 @@ Process_Upgrade(){
 		rm -f /tmp/uidivstats.sql
 		/opt/etc/init.d/S90taildns start >/dev/null 2>&1
 		touch "$SCRIPT_DIR/.upgraded"
+		touch "$SCRIPT_DIR/.upgraded2"
+	elif [ ! -f "$SCRIPT_DIR/.upgraded2" ]; then
+		/opt/etc/init.d/S90taildns stop >/dev/null 2>&1
+		{
+			echo "drop index idx_dns;"
+			echo "drop index idx_dns_clients;"
+			echo "create index idx_dns_domains on dnsqueries (ReqDmn,Timestamp);"
+			echo "create index idx_dns_time on dnsqueries (Timestamp,ReqDmn);"
+			echo "create index idx_dns_clients on dnsqueries (SrcIP,Timestamp,ReqDmn);"
+		}  > /tmp/uidivstats.sql
+		while ! "$SQLITE3_PATH" "$DNS_DB" < /tmp/uidivstats.sql >/dev/null 2>&1; do
+			sleep 1
+		done
+		rm -f /tmp/uidivstats.sql
+		/opt/etc/init.d/S90taildns start >/dev/null 2>&1
+		touch "$SCRIPT_DIR/.upgraded2"
 	fi
 }
